Added event listening to the buttons
Added checkButtonStatus which loops the buttons dictionary and starts an even listener for each button, currently only logs to the console on press event.

Also features logic to negate multiple click recordings for one click event.

Added EnterNewDirectory as a refresh function when entering a new directory but left empty for now.