Added Buttons.cs and moved logic
Added Buttons.cs and addded the logic concering a set of buttons for a specific directory to it. Designed to make it easier to keep multiple directories open at one. To open a new directory simple start a new instance with the new path.

Added logic to Buttons to set the entire set of buttons activity.

Temporarily removed search.