Fixed prompt and slight optimization in FileSelector
Previous prompt, when updated with text, would have its text display far away. Uses a button right now.

Updated for slight optimization, if the pathHistory stack is empty, just set currentDirectory to startDirectory, a stored version of the first directory, instead of generating a new version of the start directory.