Added ViewManager and "file is not.dxf" prompt
Added a ViewManager to help with the different views/screens. UIButtons in ViewManager keeps track of all the UI buttons, has add and update functions, works similarly to DirectoryManager "buttonStatus". Might integrate later.

Added functionality to display "file is not .dxf" when a file without the .dxf ending is selected, only allows for "contuine" which brings user back to main menu.